requirements.txt
# Development Requirements
pytest==6.1.1

# Production Requirements
Flask==1.1.2
opencv-python==4.4.0.42
pdf2image==1.14.0
pdfminer.six==20200726
Pillow==7.2.0
pytesseract==0.3.6
python-dotenv==0.14.0
watchdog==0.10.3
